pyproject.toml
[project]
name = "memo-chat-app-batch"
version = "0.1.0"
description = ""
requires-python = ">=3.11"
dependencies = [
    "psycopg2>=2.9.10",
    "openai>=1.12.0",
    "python-dotenv>=1.0.0",
    "stripe>=12.1.0",
    "google-generativeai>=0.8.5",
    "beautifulsoup4>=4.13.5",
    "numpy>=2.3.3",
    "lxml>=6.0.1",
]
